Change field type

+++ database/migrations/2020_04_17_172439_create_electricity_table.php
@@ -16,7 +16,7 @@ class CreateElectricityTable extends Migration
         Schema::create('electricity', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->integer('region_id');
-            $table->integer('hours');
+            $table->double('hours');
             $table->date('date');
             $table->text('comment');
             $table->string('day_period');
